pyproject.toml
[tool.poetry]
name = "ska-mid-dish-spfrx-talondx-console"
version = "0.3.9"
description = "The Mid DISH SPFRx Talon-DX Console is intended for integration and testing of the Mid DISH SPFRx hardware."
authors = [
    "Dave Del Rizzo <dave.delrizzo@nrc-cnrc.gc.ca",
    "Aditya Joshi <aditya.joshi@mda.space>",
    "Rob Huxtable <robert.huxtable@mda.space>",
    "Taylor Huang <taylor.huang@mda.space>",
]
license = "BSD license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
numpy = "^1.23.4"
jsonschema = "^4.17.0"
conan = "^1.54.0"
pytango = "^9.3.3"
matplotlib = "^3.6.2"
scipy = "^1.9.3"
polling2 = "^0.5.0"
tqdm = "^4.64.1"
parse = "^1.19.0"
argparse = "^1.4.0"
ska-tango-base = "0.10.1"
ska-ser-logging = "^0.4.1+0a677a67"
ska-ser-log-transactions = "^0.2.2+ae0f5c92"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pylint = "^2.15.5"
pylint-junit = "^0.3.2"
sphinx = "^5.1.1"
sphinx-rtd-theme = "^1.1.1"
recommonmark = "^0.7.1"
sphinx-jsonschema = "^1.19.1"
schema = "^0.7.5"
jsbeautifier = "^1.14.7"
m2r2 = "0.3.2"
sphinxcontrib-mermaid = "^0.7.1"
sphinxcontrib-plantuml = "^0.24"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-pylint = "^0.19.0"
pytest-json-report = "^1.5.0"
pytest-asyncio = "^0.20.1"

[[tool.poetry.source]]
name = "nexus-internal"
url = "https://artefact.skao.int/repository/pypi-internal/simple"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

